Add files via upload
This repository contains the Supplementary Material for the article “Multiresolution Neural Networks for Electricity Market Forecasting: Addressing Complexity with q-Wavelet Methods” (JRSS A). It includes the full Supplementary Material file illustrating the workflow and computational framework, along with the sample dataset extracted from the Australian electricity market used in the study.